+++ sdk/go/datadog/timeBoard.go
@@ -9,6 +9,8 @@ import (
 )
 
 // Provides a Datadog timeboard resource. This can be used to create and manage Datadog timeboards.
+// 
+// > **Note:**This resource is outdated. Use the new `datadog_dashboard` resource instead.
 //
 // > This content is derived from https://github.com/terraform-providers/terraform-provider-datadog/blob/master/website/docs/r/timeboard.html.markdown.
 type TimeBoard struct {