+++ sdk/nodejs/aws/integration.ts
@@ -118,6 +118,13 @@ export class Integration extends pulumi.CustomResource {
             inputs["roleName"] = args ? args.roleName : undefined;
             inputs["externalId"] = undefined /*out*/;
         }
+        if (!opts) {
+            opts = {}
+        }
+
+        if (!opts.version) {
+            opts.version = utilities.getVersion();
+        }
         super(Integration.__pulumiType, name, inputs, opts);
     }
 }